Rename component base on vue style guide
